Factor casting to separate module
